Moved _get_user_challenges back to models.py
