+++ usercontrol/views.py
@@ -4,8 +4,8 @@ from django.contrib.auth.models import User
 from django.db.models import Q
 from django.shortcuts import get_object_or_404, render, render_to_response
 from django.template import RequestContext
-from ladder.models import Rank, Challenge, Match
-from ladder.helpers import _get_user_challenges, paged
+from ladder.models import Rank, Challenge, Match, _get_user_challenges
+from ladder.helpers import paged
 from datetime import datetime
 
 PROFILE_RECENT_MATCHES    = 5         # How many matches to show under the "Recent Matches" header
